# bump `go-licenses` to v2.0.0-alpha.1