--- a/ci/overlay.nix
+++ b/ci/overlay.nix
@@ -5,6 +5,7 @@
   controller-gen = pkgs.callPackage ./controller-gen.nix { };
   crd-ref-docs = pkgs.callPackage ./crd-ref-docs.nix { };
   docker-tag-list = pkgs.callPackage ./docker-tag-list.nix { };
+  go-licenses = pkgs.callPackage ./go-licenses.nix { inherit prev; };
   helm-3-10-3 = pkgs.callPackage ./helm.nix { };
   kuttl = pkgs.callPackage ./kuttl.nix { };
   setup-envtest = pkgs.callPackage ./setup-envtest.nix { };