This commit bumps the `go-licenses` tool to the latest pre-release which contains significant performance improvements and is _hopefully_ more deterministic as there have been an increasing number of flakes in CI as of late.

Before:
```
Executed in   37.28 secs    fish           external
   usr time   43.05 secs    0.12 millis   43.05 secs
   sys time    1.69 secs    1.89 millis    1.69 secs
```

After:
```
Executed in   10.08 secs    fish           external
   usr time    5.14 secs   62.00 micros    5.14 secs
   sys time    1.13 secs  695.00 micros    1.12 secs
```